WebDevAssignment4/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1

Django==4.1
djangorestframework==3.14
gunicorn==20.1
psycopg2==2.9  # If using PostgreSQL
channels==4.0  # If using Django Channels
channels_redis==3.0  # For Redis-based Channels layer (optional)
